Revert "WicketWICKET-6334 WicketObjects#sizeof() should detach Sessions"
This reverts commit 8b6fcd869ceb96f7b4ea003d3d5665a1626390ad.
