+++ wicket-core/src/test/java/org/apache/wicket/serialize/java/JavaSerializerTest.java
@@ -29,7 +29,11 @@ import java.io.Serializable;
 
 import org.apache.wicket.WicketRuntimeException;
 import org.apache.wicket.core.util.objects.checker.CheckingObjectOutputStream;
+import org.apache.wicket.core.util.objects.checker.IObjectChecker;
+import org.apache.wicket.core.util.objects.checker.NotDetachedModelChecker;
 import org.apache.wicket.markup.html.WebComponent;
+import org.apache.wicket.model.IModel;
+import org.apache.wicket.model.LoadableDetachableModel;
 import org.apache.wicket.util.io.IOUtils;
 import org.apache.wicket.util.tester.WicketTestCase;
 import org.junit.Test;
@@ -39,6 +43,45 @@ import org.junit.Test;
  */
 public class JavaSerializerTest extends WicketTestCase
 {
+	/**
+	 * https://issues.apache.org/jira/browse/WICKET-4812
+	 *
+	 * Tests that the serialization fails when a checking ObjectOutputStream is
+	 * used with NotDetachedModelChecker and there is a non-detached LoadableDetachableModel
+	 * in the object tree.
+	 */
+	@Test
+	public void notDetachedModel()
+	{
+		JavaSerializer serializer = new JavaSerializer("JavaSerializerTest")
+		{
+			@Override
+			protected ObjectOutputStream newObjectOutputStream(OutputStream out) throws IOException
+			{
+				IObjectChecker checker = new NotDetachedModelChecker();
+				return new CheckingObjectOutputStream(out, checker);
+			}
+		};
+
+		IModel<String> model = new NotDetachedModel();
+		model.getObject();
+		WebComponent component = new WebComponent("id", model);
+		byte[] serialized = serializer.serialize(component);
+		assertNull("The produced byte[] must be null if there was an error", serialized);
+	}
+
+	/**
+	 * A Model used for #notDetachedModel() test
+	 */
+	private static class NotDetachedModel extends LoadableDetachableModel<String>
+	{
+		@Override
+		protected String load()
+		{
+			return "loaded";
+		}
+	}
+
 	/**
 	 * https://issues.apache.org/jira/browse/WICKET-4812
 	 *
